requirements.txt
transformers>=4.41.0
accelerate>=0.32.0
torch>=2.2.0
datasets>=2.19.0
pyyaml>=6.0.1
numpy>=1.24.0
tqdm>=4.66.0
rich>=13.7.0
jinja2>=3.1.4
